Revert code and dependencies for Java 8 compatibility
Dependencies include:

* Selenium 4.10.0 -> 4.4.0
* Javalin 5.6.1 -> 4.6.8
* Logback 1.4.8 to 1.3.11

Additionally, in the `WdmServer.java` file, the type of variables `requestMethod` has been changed from an enum to a string. The method signature of the `exchange()` method in the same file has also been modified to accept a string parameter instead of an enum.

Finally, in the `WebDriverManager.java` file, an argument is added at the beginning of the array before calling a method to accommodate 4.x style CLI syntax.